[plugin.video.metalvideo] 3.0.7
