+++ plugin.video.metalvideo/addon.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.metalvideo" name="Metalvideo" provider-name="willforde" version="3.0.5">
+<addon id="plugin.video.metalvideo" name="Metalvideo" provider-name="willforde" version="3.0.7">
     <requires>
         <import addon="xbmc.python" version="2.25.0"/>
-        <import addon="script.module.codequick" version="0.9.2"/>
+        <import addon="script.module.codequick" version="0.9.6"/>
         <import addon="plugin.video.youtube" version="5.0.0"/>
     </requires>
     <extension point="xbmc.python.pluginsource" library="addon.py">
@@ -25,6 +25,7 @@
             <screenshot>resources/screenshots/screenshot001.jpg</screenshot>
             <screenshot>resources/screenshots/screenshot002.jpg</screenshot>
         </assets>
-        <news>[fix] Fixed playback problem with Kodi on Xbox.</news>
+        <news>[upd] Changed to work with the updated site.</news>
+        <reuselanguageinvoker>true</reuselanguageinvoker>
     </extension>
 </addon>
